refactor(mespapiers): Split content from queries in Home
no code change
